Move common test classes
