+++ instrumentation/rxjava/rxjava-3-common/testing/build.gradle.kts
@@ -7,8 +7,8 @@ dependencies {
 
   api("io.reactivex.rxjava3:rxjava:3.0.12")
 
+  implementation("io.opentelemetry:opentelemetry-extension-annotations")
   implementation("com.google.guava:guava")
-
   implementation("org.codehaus.groovy:groovy-all")
   implementation("io.opentelemetry:opentelemetry-api")
   implementation("org.spockframework:spock-core")
