+++ instrumentation/rxjava/rxjava-3.1.1/javaagent/src/test/groovy/RxJava3SubscriptionTest.groovy
@@ -7,5 +7,4 @@ import io.opentelemetry.instrumentation.rxjava3.AbstractRxJava3SubscriptionTest
 import io.opentelemetry.instrumentation.test.AgentTestTrait
 
 class RxJava3SubscriptionTest extends AbstractRxJava3SubscriptionTest implements AgentTestTrait {
-
 }
